+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Implement support for the `sushi-config.yaml` file: annotations, auto-completion and validation of properties
+  (when possible)
+
 ## [0.0.6] - 2025-12-21
 
 ### Added